requirements.txt
aiohttp==3.9.3
asyncio==3.4.3
pillow==10.1.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0 